LayoutTests/fast/dom/popup-menu-client-test.html

<html>
  <head>
    <body>
      <div id="console"></div>
      <select id="selectelement">
        <option>Element 1</option>
        <option>Element 2</option>
        <option>Element 3</option>
      </select>
    </body>
    <script src="../js/resources/js-test-pre.js"></script>
    <script>
      var selectElement = document.getElementById("selectelement");
      if (window.internals) {
        // FIXME: This wil fail if the page scale factor != 1.0
        var pageScaleFactor = 2.0;
        eventSender.scalePageBy(pageScaleFactor, 0, 0);

        var popupClientPaddingLeft = window.internals.popupClientPaddingLeft(selectElement);
        var popupClientPaddingRight = window.internals.popupClientPaddingRight(selectElement);
        var boundingBoxRect = window.internals.popupClientBoundingBoxRect(selectElement);
        var boundingClientRect = selectElement.getBoundingClientRect();
        shouldBe("popupClientPaddingLeft", "4");
        shouldBe("popupClientPaddingRight", "2");
        shouldBe("boundingBoxRect.left", "boundingClientRect.left");
        shouldBe("boundingBoxRect.top", "boundingClientRect.top");
        shouldBe("boundingBoxRect.width", "boundingClientRect.width");
        shouldBe("boundingBoxRect.height", "boundingClientRect.height");
      }
      if (window.layoutTestController) {
        layoutTestController.dumpAsText();
      }
      successfullyParsed = true;
    </script>
    <script src="../js/resources/js-test-post.js"></script>
    </script>
  </head>
</html>
